move load_anat and MNI152Template to plotting.image._utils
