+++ nilearn/plotting/html_stat_map.py
@@ -22,7 +22,7 @@ from nilearn._utils.param_validation import check_threshold
 from nilearn.datasets import load_mni152_template
 from nilearn.image import get_data, new_img_like, reorder_img, resample_to_img
 from nilearn.plotting.find_cuts import find_xyz_cut_coords
-from nilearn.plotting.image.img_plotting import load_anat
+from nilearn.plotting.image import load_anat
 from nilearn.plotting.js_plotting_utils import colorscale, get_html_template
