+++ nilearn/plotting/image/__init__.py
@@ -1,3 +1,4 @@
+from nilearn.plotting.image._utils import load_anat
 from nilearn.plotting.image.img_plotting import (
     plot_anat,
     plot_carpet,
@@ -13,6 +14,7 @@ from nilearn.plotting.image.img_plotting import (
 )
 
 __all__ = [
+    "load_anat",
     "plot_anat",
     "plot_carpet",
     "plot_connectome",
